refactor(signals): move patchState to separate file
