+++ modules/signals/src/signal-state.ts
@@ -2,14 +2,14 @@ import { signal, WritableSignal } from '@angular/core';
 import { DeepSignal, toDeepSignal } from './deep-signal';
 import { defaultEqual } from './select-signal';
 
-type SignalState<State extends Record<string, unknown>> = DeepSignal<State> &
-  SignalStateMeta<State>;
+export const SIGNAL_STATE_META_KEY = Symbol('SIGNAL_STATE_META_KEY');
 
-const SIGNAL_STATE_META_KEY = Symbol('SIGNAL_STATE_META_KEY');
-
-type SignalStateMeta<State extends Record<string, unknown>> = {
+export type SignalStateMeta<State extends Record<string, unknown>> = {
   [SIGNAL_STATE_META_KEY]: WritableSignal<State>;
 };
+
+type SignalState<State extends Record<string, unknown>> = DeepSignal<State> &
+  SignalStateMeta<State>;
 
 /**
  * Signal state cannot contain optional properties.
@@ -19,10 +19,6 @@ type NotAllowedStateCheck<State> = State extends Required<State>
     ? { [K in keyof State]: State[K] & NotAllowedStateCheck<State[K]> }
     : unknown
   : never;
-
-export type PartialStateUpdater<State extends Record<string, unknown>> =
-  | Partial<State>
-  | ((state: State) => Partial<State>);
 
 export function signalState<State extends Record<string, unknown>>(
   initialState: State & NotAllowedStateCheck<State>
@@ -34,19 +30,4 @@ export function signalState<State extends Record<string, unknown>>(
   });
 
   return deepSignal as SignalState<State>;
-}
-
-export function patchState<State extends Record<string, unknown>>(
-  signalState: SignalStateMeta<State>,
-  ...updaters: PartialStateUpdater<State>[]
-): void {
-  signalState[SIGNAL_STATE_META_KEY].update((currentState) =>
-    updaters.reduce(
-      (nextState: State, updater) => ({
-        ...nextState,
-        ...(typeof updater === 'function' ? updater(nextState) : updater),
-      }),
-      currentState
-    )
-  );
 }
